Refactor Challenge Service to reduce code complexity
