+++ src/main/java/com/techlooper/service/ChallengeService.java
@@ -2,12 +2,10 @@ package com.techlooper.service;
 
 import com.techlooper.entity.ChallengeEntity;
 import com.techlooper.entity.ChallengeRegistrantDto;
-import com.techlooper.entity.ChallengeRegistrantEntity;
-import com.techlooper.model.*;
-import freemarker.template.TemplateException;
+import com.techlooper.model.ChallengeDetailDto;
+import com.techlooper.model.ChallengeDto;
+import com.techlooper.model.ChallengePhaseEnum;
 
-import javax.mail.MessagingException;
-import java.io.IOException;
 import java.util.List;
 
 /**
@@ -15,28 +13,11 @@ import java.util.List;
  */
 public interface ChallengeService {
 
-    ChallengeEntity savePostChallenge(ChallengeDto challengeDto) throws Exception;
+    ChallengeEntity postChallenge(ChallengeDto challengeDto);
 
-    void sendPostChallengeEmailToEmployer(ChallengeEntity challengeEntity)
-            throws MessagingException, IOException, TemplateException;
-
-    void sendPostChallengeEmailToTechloopies(ChallengeEntity challengeEntity, Boolean isNewChallenge)
-            throws MessagingException, IOException, TemplateException;
-
-    void sendEmailNotifyRegistrantAboutChallengeTimeline(ChallengeEntity challengeEntity,
-                                                         ChallengeRegistrantEntity challengeRegistrantEntity, ChallengePhaseEnum challengePhase, boolean isSpecificDayNotification) throws Exception;
-
-    void sendEmailNotifyEmployerWhenPhaseClosed(ChallengeEntity challengeEntity, ChallengePhaseEnum currentPhase,
-                                                ChallengePhaseEnum oldPhase) throws Exception;
+    Long joinChallenge(ChallengeRegistrantDto challengeRegistrantDto);
 
     ChallengeDetailDto getChallengeDetail(Long challengeId, String loginEmail);
-
-    void sendApplicationEmailToContestant(ChallengeEntity challengeEntity, ChallengeRegistrantEntity challengeRegistrantEntity)
-            throws MessagingException, IOException, TemplateException;
-
-    ChallengeRegistrantEntity joinChallengeEntity(ChallengeRegistrantDto challengeRegistrantDto);
-
-    long joinChallenge(ChallengeRegistrantDto challengeRegistrantDto);
 
     List<ChallengeDetailDto> listChallenges();
 
@@ -50,23 +31,11 @@ public interface ChallengeService {
 
     ChallengeDetailDto getTheLatestChallenge();
 
-    boolean delete(Long id, String ownerEmail);
-
-    ChallengeDto findChallengeById(Long id, String ownerEmail);
-
-    void sendDailySummaryEmailToChallengeOwner(ChallengeEntity challengeEntity) throws Exception;
+    boolean deleteChallenge(Long challengeId, String ownerEmail);
 
     boolean isOwnerOfChallenge(String ownerEmail, Long challengeId);
 
-    ChallengeEntity findChallengeIdAndOwnerEmail(Long challengeId, String ownerEmail);
-
-    boolean sendEmailToDailyChallengeRegistrants(String challengeOwner, Long challengeId, Long now, EmailContent emailContent);
-
-    boolean sendEmailToRegistrant(String challengeOwner, Long registrantId, EmailContent emailContent);
-
-    void updateSendEmailToContestantResultCode(ChallengeRegistrantEntity challengeRegistrantEntity, EmailSentResultEnum code);
-
-    void updateSendEmailToChallengeOwnerResultCode(ChallengeEntity challengeEntity, EmailSentResultEnum code);
+    ChallengeEntity findChallengeById(Long challengeId, String ownerEmail);
 
     void calculateChallengePhases(ChallengeDetailDto challengeDetailDto);
