+++ src/main/java/com/techlooper/util/DateTimeUtils.java
@@ -2,10 +2,8 @@ package com.techlooper.util;
 
 import org.joda.time.DateTime;
 import org.joda.time.Days;
-import org.joda.time.format.DateTimeFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.util.StringUtils;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -22,28 +20,10 @@ public class DateTimeUtils {
 
     public static final String BASIC_DATE_TIME_PATTERN = "dd/MM/yyyy HH:mm";
 
-//  public static final String ISO_DATETIME_PATTERN = "yyyy-MM-dd'T'HH:mm:ssZZ";
-
-    public static DateTime parseBasicDate(String text) {
-        if (StringUtils.hasText(text)) {
-            try {
-                return DateTimeFormat.forPattern(BASIC_DATE_PATTERN).parseDateTime(text);
-            } catch (Exception e) {
-                LOGGER.debug("Couldn't parse date-time [{}] by format [{}]", text, BASIC_DATE_PATTERN);
-            }
-        }
-        return null;
-    }
-
     public static Date string2Date(String datetime, String pattern) throws ParseException {
         SimpleDateFormat formatter = new SimpleDateFormat(pattern);
         return formatter.parse(datetime);
     }
-
-//    public static String date2String(Date datetime, String pattern) {
-//        SimpleDateFormat formatter = new SimpleDateFormat(pattern);
-//        return formatter.format(datetime);
-//    }
 
     public static String currentDate() {
         SimpleDateFormat formatter = new SimpleDateFormat(BASIC_DATE_PATTERN);
@@ -67,6 +47,10 @@ public class DateTimeUtils {
         return formatter.format(new Date());
     }
 
+    public static Long currentDateTime() {
+        return new Date().getTime();
+    }
+
     public static int daysBetween(Date firstDate, Date secondDate) {
         return Days.daysBetween(new DateTime(firstDate), new DateTime(secondDate)).getDays();
     }
@@ -75,10 +59,6 @@ public class DateTimeUtils {
         DateTime firstDateTime = new DateTime(string2Date(firstDateStr, BASIC_DATE_PATTERN));
         DateTime secondDateTime = new DateTime(string2Date(secondDateStr, BASIC_DATE_PATTERN));
         return Days.daysBetween(firstDateTime, secondDateTime).getDays();
-    }
-
-    public static void main(String args[]) {
-        System.out.println(yesterdayDate());
     }
 
 }
